pyproject.toml
[project]
name = "pig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bitarray>=3.4.2",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "pyside6>=6.9.0",
    "pytest>=8.3.5",
    "ruff>=0.11.12",
    "scipy>=1.15.3",
    "tqdm>=4.67.1",
]

[tool.uv.workspace]
members = ["bla"]

[tool.ruff.lint]
select = ["I"]
ignore = ["F401"]

[dependency-groups]
dev = [
    "snakeviz>=2.2.2",
    "tuna>=0.5.11",
]
